backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.1
requests==2.32.2
python-docx==1.1.2
requests==2.32.2
playwright
